TTD implementation and bug fix
Minor prefabs now must be hovered for a while for the description to appear. The time for that to happen (TimeToDescription) can be set in the MinorArcanaPrefab.

Fixed a bug where if the mouse stopped hovering on a card while the game was in pause the description wouldn't close itself